+++ src/main/java/de/berg/systeme/jenkins/wix/ToolsetException.java
@@ -1,5 +1,7 @@
 package de.berg.systeme.jenkins.wix;
 
+import java.util.ResourceBundle;
+
 /***
  * Thrown in case of an exception while executing the {@link Toolset}.
  * 
@@ -9,6 +11,7 @@ package de.berg.systeme.jenkins.wix;
  */
 public class ToolsetException extends Throwable {
 	private static final long serialVersionUID = 7128147634196518698L;
+    private static final ResourceBundle messages = ResourceBundle.getBundle("Messages");
 
 	public ToolsetException(String message) {
 		super(message);